requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-admin-cli==0.1.1
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
